Fix health checks and restore test compatibility
