+++ admin/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+bp = Blueprint("admin", __name__)
+
+@bp.route("/")
+def index():
+    return "Admin Portal placeholder"
