+++ api/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint, jsonify
+
+bp = Blueprint("api", __name__)
+
+@bp.route("/health")
+def health():
+    return jsonify({"ok": True})
